Source/Minor: Build separate thumbnail view
The UI has been redesigned with a new idea of
thumbnails. Each shots taken are stored
in the `USER_CACHE_DIR` which resolves
to ~/.cache on debian. These are loaded into
a viewport which displays a tiny thumbnail.
Any of the entries in the side view when
clicked, expands into a bigger view pane.

Upcoming:
    - Implement a smarter expansion mechanism
      for the side view. Currently, the
      view expands in an horrible manner.
    - Improve hacks in code.

Signed-off-by: Meekness Adesina <zenon8adams@gmail.com>