Source/feat: Implement page rotation
The orientation of pages can be changed
by pressing Ctrl+R for clockwise rotation
and Ctrl+L for anti-clockwise rotation.
Each orientation keeps track of the
current panning state. Panning to position
before rotation restores the panned state
on returning to the same orientation.

Setup rotation animation for the settings
button giving live to the MainViewer.

The MainViewer has been configured to be
constrained by the position of the gnome
panel. The `Expand` indicator has been
modified to follow suite.

Signed-off-by: Meekness Adesina <zenon8adams@gmail.com>