Source/Fix: Resolve sticky cursor on drag
If window is in drag mode and the cursor gets captured
by a button, vfunc_key_release_event will not be
triggered. The old implementation depends on this
event to be fired before releasing the cursor. To fix
this, we listen to enter-event on all buttons and
release the cursor on receiving such event.

Window tiling mode is disabled due to implementation
flaw. If would certainly be fixed in the next release.

Signed-off-by: Meekness Adesina <zenon8adams@gmail.com>