Source/Fix: Window vertical expansion glitch
When the window is dragged vertically and it hits
the top and bottom boundary, it tends to bounce
irregularly around that boundary giving off visual
artifacts. To fix this, we use only update some
endpoints based on if we are in drag mode or resize
mode.

Remove registered signal handlers for the `shutter`.

Signed-off-by: Meekness Adesina <zenon8adams@gmail.com>