Source/Minor: Improve startup time for pixzzle
The time it takes for the shot tool to be activated
has been improved by implementing thumbnail cache.
The caching system saves a thumbnail image the
first time the image is loaded into the
`ThumbnailViewer`. This is the file/image that is
loaded at every other time.

We also save an indicator as an ancilliary chunk
in the png screenshot that assists us in determining
which thumbnail ought to be blurred. Thumbnails
smaller in size than the `ThumbnailViewer`'s size
are always blurred.

Signed-off-by: Meekness Adesina <zenon8adams@gmail.com>