Source/feat: Implement text extraction via API
When the tool is open, pressing the shortcut `O` will
trigger the OCR tool changing the cursor to CROSSHAIR.
A box can be craved out by dragging the cursor around
a text to be extracted. Once this is done, a loading
indicator pops out to indicate that the ocr processing
is taking place. This is with the aid of external ocr
api called `ocr-space`. Open Source ocr tools are not
currently well refined. They are also too heavy and
require external tools to be able to configure locally.
Once the processing is done at `ocr-space`, the
resulting text from the extraction (if successful) is
sent back which replaces the content of the loading
indicator. `Ctrl+C` or by clicking on the loading
indicator copies out the extracted text into the
clipboard. The shortcut `O` is a toggle trigger,
activating it again moves the application out of ocr
mode.

Signed-off-by: Meekness Adesina <zenon8adams@gmail.com>