Source/Minor: Improve OCR user experience
The ocrtip window vibrates and change color
if an error occurs when processing request.

The `Esc` (escape) key can be pressed to
cancel an ongoing ocr session at anytime of
the processing.

Minor code refactoring by use of constant
in place of magic numbers especially in
indicating session status codes.

Signed-off-by: Meekness Adesina <zenon8adams@gmail.com>