Source/Minor: Code refactoring and enhancement
We use wrapper class to note when cursor enters
into reactive widgets and implement
`vfunc_motion_event` setting up the cursor to
the desired cursor instead of handling it
centrally through the use of
`_isInRestrictedArea` method.

Use of layoutManagers `addChrome` inplace of
`addTopChrome` as previous method places item
in dnd (drag and drop) behind the viewer which
shouldn't be.

Signed-off-by: Meekness Adesina <zenon8adams@gmail.com>