Source/feat: Perform OCR on snip section
When the shutter is opened, the `showPointer`
indicator has been replaced with a more
functional `performOCR` indicator. When
this button is toggled, OCR is performed
on the entire snip area immediately the shot
is taken.

The OCR result textbox has been constrained
with the use of vfunc_allocate hook. This
prevents it from blowing out of proportion
and covering up buttons on the screen.

Signed-off-by: Meekness Adesina <zenon8adams@gmail.com>