Source/feat: Implement group-by-date sidebar
A swap button has been designed and placed
beside the `Add New` button and just
underneath the `ThumbnailViewer`. When this
button is toggled, it flips between the
`ThumbnailViewer` view of the shots and
the `FolderViewer` view.

A common base class `UISideViewBase` now
holds the scaffolding that defines what a
scrollable sidebar ought to have.
`UIThumbnailViewer` and `UIFolderViewer`
entends this base class.

The `Up` and `Down` arrow keys have been
bound to serve as shortcuts for the swap
button.

The `FolderViewer` holds a grouping of
the all shots taken on the same date
and adorn it with a gradient defined
in the `objects/gradients.json` file.

Add tool tip to the settings and swap
buttons to improve user experience.
The `Tooltip` has been extracted into
its own component file.

Cached thumbnails saved in the `.thumbnail`
folder are not deleted when the corresponding
image is deleted. This has been fixed.

Signed-off-by: Meekness Adesina <zenon8adams@gmail.com>