Source/fix: Expanding window doesn't update cursor
During window resize, the cursor changes to indicate
the expansion direction. The bug introduced
in @71b2c8d causes the cursor to update when there
is a paint event in the `UIImageRenderer` canvas.
The cursor update function called is the related
to changing the cursor from `snip` mode to `default`
mode. This effect subsumes the cursor update function
from `UIMainViewer` that is responsible for changing
the cursor on expanding the window.

Signed-off-by: Meekness Adesina <zenon8adams@gmail.com>