Source/feat: Add group reset and dynamic loader
A group reset button has been added to retore the
ordering of screenshots to the flat model where
all shots are not grouped.

`St.ScrollView` tries to allocate more space than
required when large contents are pushed into it.
This causes a glitch in the application where
the `ScrollView` stretches past the boundary of
its parent windows.
To fix this, we calculate the
available container size and load only into the
visible slots. We then use a dynamic loader to
load the remaining gradually.
The dynamic loader is controlled to pause itself
when its parent becomes hidden and resume its
operation when the parent is staged.

The group reset button `meltButton` is an animated
button. We use a sprite sheet to achive this and
let `Animation` class handle the frame rendering
for us.

Lastly, the `meltButton` is synced with the
the side view toggle button (`swapButton`) to give
gives a better user experience.

Signed-off-by: Meekness Adesina <zenon8adams@gmail.com>