Source/fix: Fix empty image viewer on deletion
If a picture belonging to a randomly selected
group is deleted and that group becomes empty,
if the user deletes all pictures one after the
other, upon reaching that group where all
pictures have been emptied, the image viewer
will not display any picture since the current
group in view has become empty. To fix this,
when `UIThumbnailViewer` triggers a reload,
we delete all empty groups with the function
`_clearEmptyGroups`. With this, we are
guaranteed of a free sail over this hole.

Some Widgets have either been misparented or
parented twice. This have been removed.

Signed-off-by: Meekness Adesina <zenon8adams@gmail.com>