Source/fix: Remove batch loading hack
Due to the `crossSlideAnimation` extending the
height of its participants, this height
automatically becomes the max height of such
widget. If its parent container shrinks lesser
than that height, the widget overflows it
parent. To solve this, we hook into `_updateSize`
`UIMainViewer` function and compute the height
growth distance. We use this distance to update
the height of the currently visible widget.

Signed-off-by: Meekness Adesina <zenon8adams@gmail.com>