Source/support: Drop support for Tiling
The ability to expand the views past its
adaptive size is termed tiling. It was
still in development phase before being
deemed as unnecessary. Users can always
expand the the viewer to be able to see
more or pan around the picture.

The response time of dragging the view
has been improved significantly. The view
responds well to drag action and does not
lag as before.

Signed-off-by: Meekness Adesina <zenon8adams@gmail.com>