Source/feat: Enable/Disable apps on request
Applications in the dock can how show indicators
for when they are disabled/enabled. We convert
the app icon into a grayscale image which we
then use as the disabled icon. Apps have a
`can_disable()` function which can be queried
to determine if they can be disabled.

When applications with animatable icons become
disabled, their animation stops playing. This
leaves an blank icon inplace of the still icon
of the app. To fix this, we the first slice
from the animated icon sprite and use this as
our still icon.

View composition has been greatly simplified with
the use of constraints. Arrangement of elements
have been switched from the previouse nested
`UILayout` implementation into a flat model which
uses constraints.

Keystrokes mapped to `disabled` applications will
now notify the user to for example, add a new
screenshot to the empty stage.

The `crossSlideAnimation` has been improved through
the use of `clip_to_allocation` flag. With this,
we don't have to compute height of the views. We
can just 'ease' to any height. The parent viewport
will clip to its own height.

Signed-off-by: Meekness Adesina <zenon8adams@gmail.com>