Source/fix: Crash report when viewer is relaunched
Applications were previously managed by two different
widgets which requires detach from the two widgets
upon destruction. When the viewer is destroyed, one
of the widget destroys its child while the other
is left hanging. Then the app is relaunched, the
hanging app is still being held. This causes crash
when the app is clicked. The fix to this is to remove
one of the reference being managed by one of the
widgets.

In `_bindConstraints`, we use a tag `_bindDir` to
prevent binding to the same side twice. Without this
we introduce errors in the layout of the widgets.

Cleanup of all used widget have been improved.

Signed-off-by: Meekness Adesina <zenon8adams@gmail.com>