Source/Minor: Allow separate disable/enable of app
Applications can now be independently enabled or
disabled by passing in an array of apps that is
intended to be disabled/enabled.

UI glitches was erroneously introduced by previous
commit @0261a405. The issue occurred when we try
to destroy all appIcons `iconContainer` children
instead of removing the offending single child
`icon`.

Signed-off-by: Meekness Adesina <zenon8adams@gmail.com>