Source/Fix: Improve view drag/keypress experience
Dragging the viewer lags behind the cursor. This
is to the `vfunc_enter_event` acquiring `key focus`
on cursor entering the viewer and `releasing it
in `vfunc_leave_event`. This is fixed by acquiring
the key focus if it is not already held and
releasing the focus when the mouse is outside the
viewer and a key is pressed.

When the settings button is clicked. It may be
showned in such a way that it is occluded by the
viewer. We fix this by moving the viewer to its
default position and then moving the settings
window to the center, if there's significant
overlap between them.

Signed-off-by: Meekness Adesina <zenon8adams@gmail.com>