Source/Minor: Add expand button to screenshot tool
Sometimes, it is easier to take to start dragging
the screenshot frame handles from a position where
it covers the whole screen. For example, if a user
wants to take screenshot of the work area (area
excluding the panels and dock) it is better to
have the screenshot tool span the whole screen
instead of dragging the handles all the way to
the edges of the monitor. This button does this
expansion.

Signed-off-by: Meekness Adesina <zenon8adams@gmail.com>