Source/Feat: Implement zoom mode on image viewer
This draft contains the initial design of zoomable
image viewer.
This implementation also includes able to pan around
the zoomed viewer. The design principle surrounding
this implementation is that zooming in an image should
not make the viewer resize past its original size.
Zooming out to any extent doesn't grow the image by the
zoom factor. We zoom only the visible part of the image.
The maximum we will ever scale to is the size of the
viewer. This allows us to be able to zoom to any extent
without hitting performance penalty.
Also, zooming is handled as though we are zooming the
image with the center of the image as the axis even if
the center is not visible in the viewer.

When in the zoom mode, we use different set of variables
to control the effects of zoom and synchronize activities
at zoom levels with zoom level 1.
when in this "zoomed" mode, resizing the viewer adapts
the contained image as it does for zoom level 1.

Updates are made to other parts of the application to make
sure that the addition of zoom doesn't break features added
earlier.

Signed-off-by: Meekness Adesina <zenon8adams@gmail.com>