Source/Fix: Inaccurate zoom for image wrt viewport
When images are smaller than the viewport, the
current implementation zooms the image with
respect to its current size. It doesn't zoom
a small image to cover the size of the viewport.
It zooms the image to cover the space taken by
the image.
This implementation fixes this issue by first
updating the expansion points of the image,
scaling the image, then apply panning to clip
off the invisible sections of the image.

`connectObject` is used to join multiple signals
together and to also remove signals when a
registered object is destroyed.

Signed-off-by: Meekness Adesina <zenon8adams@gmail.com>