Source/Minor: Disable modal watch on viewer hidden
When the viewer is unstaged, the watcher for the
modal is temporarily masked such that it handler
is called only when the watcher becomes visible.

Minor bug fixes for zoom level implementation for
image fitting into viewport.

`_redisplay` function is triggered more than once,
we want a behaviour where it is executed only when
its controller is mapped also. We check that the
controller is mapped before taking any action.

Signed-off-by: Meekness Adesina <zenon8adams@gmail.com>